Added Base Character Model
The white background is a placeholder, I can make it transparent. This is inteded for testing purposes and will be used as a base for further characters. If there is anything you want me to change or add let me know. Files are in PNG format and 18x26. Naming convention is the type of image, type of action, direction, and whether it's the first or second frame.